requirements.txt
aniso8601==9.0.1
attrs==21.2.0
click==8.0.3
Flask==2.1.2
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
jsonschema==4.1.2
MarkupSafe==2.1.1
marshmallow==3.14.0
pyrsistent==0.18.0
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.39
typing-extensions==3.10.0.2
Werkzeug==2.1.2
zipp==3.6.0

PyJWT~=2.4.0